adding a wrapper component
